Moves move dialogs into a dialogs folder
