+++ src/components/dialogs/AngelSpecialDialog.tsx
@@ -1,25 +1,27 @@
 import React, { FC, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useMutation } from '@apollo/client';
-import { Box, FormField, Select, TextInput } from 'grommet';
+import { Box, Select } from 'grommet';
 
-import DialogWrapper from './DialogWrapper';
-import { StyledMarkdown } from './styledComponents';
-import { HeadingWS, ParagraphWS, ButtonWS, RedBox, healHarmBackground } from '../config/grommetConfig';
-import PERFORM_HEAL_HARM_MOVE, { PerformHealHarmMoveData, PerformHealHarmMoveVars } from '../mutations/performHealHarmMove';
-import { CharacterMove, Move } from '../@types/staticDataInterfaces';
-import { useFonts } from '../contexts/fontContext';
-import { useGame } from '../contexts/gameContext';
+import DialogWrapper from '../DialogWrapper';
+import { StyledMarkdown } from '../styledComponents';
+import { HeadingWS, ParagraphWS, ButtonWS, angelSpecialBackground } from '../../config/grommetConfig';
+import PERFORM_ANGEL_SPECIAL_MOVE, {
+  PerformAngelSpecialMoveData,
+  PerformAngelSpecialMoveVars,
+} from '../../mutations/performAngelSpecialMove';
+import { CharacterMove, Move } from '../../@types/staticDataInterfaces';
+import { useFonts } from '../../contexts/fontContext';
+import { useGame } from '../../contexts/gameContext';
 
-interface HealHarmDialogProps {
+interface AngelSpecialDialogProps {
   move: Move | CharacterMove;
   handleClose: () => void;
 }
 
-const HealHarmDialog: FC<HealHarmDialogProps> = ({ move, handleClose }) => {
+const AngelSpecialDialog: FC<AngelSpecialDialogProps> = ({ move, handleClose }) => {
   // -------------------------------------------------- Component state ---------------------------------------------------- //
   const [otherCharacterId, setotherCharacterId] = useState('');
-  const [harm, setHarm] = useState(0);
   // -------------------------------------------------- 3rd party hooks ---------------------------------------------------- //
   const { gameId } = useParams<{ gameId: string }>();
 
@@ -28,21 +30,15 @@ const HealHarmDialog: FC<HealHarmDialogProps> = ({ move, handleClose }) => {
   const { userGameRole, otherPlayerGameRoles } = useGame();
 
   // ------------------------------------------------------ graphQL -------------------------------------------------------- //
-  const [performHealHarmMove, { loading: performingHealHarmMove }] = useMutation<
-    PerformHealHarmMoveData,
-    PerformHealHarmMoveVars
-  >(PERFORM_HEAL_HARM_MOVE);
+  const [performAngelSpecialMove, { loading: performingAngelSpecialMove }] = useMutation<
+    PerformAngelSpecialMoveData,
+    PerformAngelSpecialMoveVars
+  >(PERFORM_ANGEL_SPECIAL_MOVE);
 
   // ------------------------------------------------- Component functions -------------------------------------------------- //
   const characters = otherPlayerGameRoles?.map((gameRole) => gameRole.characters[0]) || [];
-  const handleHealHarmMove = () => {
-    if (
-      !!userGameRole &&
-      userGameRole.characters.length === 1 &&
-      !performingHealHarmMove &&
-      harm > 0 &&
-      !!otherCharacterId
-    ) {
+  const handleAngelSpecialMove = () => {
+    if (!!userGameRole && userGameRole.characters.length === 1 && !performingAngelSpecialMove && !!otherCharacterId) {
       const otherGameroleId = otherPlayerGameRoles?.find((gameRole) => {
         let isMatch = false;
         gameRole.characters.forEach((character) => {
@@ -54,14 +50,13 @@ const HealHarmDialog: FC<HealHarmDialogProps> = ({ move, handleClose }) => {
       if (!otherGameroleId) return;
 
       try {
-        performHealHarmMove({
+        performAngelSpecialMove({
           variables: {
             gameId,
             gameroleId: userGameRole.id,
             otherGameroleId,
             characterId: userGameRole.characters[0].id,
             otherCharacterId,
-            harm,
           },
         });
         handleClose();
@@ -74,40 +69,24 @@ const HealHarmDialog: FC<HealHarmDialogProps> = ({ move, handleClose }) => {
   // ------------------------------------------------------ Render -------------------------------------------------------- //
 
   return (
-    <DialogWrapper background={healHarmBackground} handleClose={handleClose}>
-      <Box gap="24px">
+    <DialogWrapper background={angelSpecialBackground} handleClose={handleClose}>
+      <Box gap="12px">
         <HeadingWS crustReady={crustReady} level={4} alignSelf="start">
           {move.name}
         </HeadingWS>
         <StyledMarkdown>{move.description}</StyledMarkdown>
-        <Box fill direction="row" align="start" justify="between" pad="12px" gap="12px">
-          <Box fill>
-            <ParagraphWS alignSelf="start">Who did you heal?</ParagraphWS>
-            <Select
-              id="target-character-input"
-              aria-label="target-character-input"
-              name="target-character"
-              placeholder="Who?"
-              options={characters}
-              labelKey={'name'}
-              valueKey={'id'}
-              onChange={(e) => setotherCharacterId(e.value.id)}
-            />
-          </Box>
-          <Box fill>
-            <ParagraphWS alignSelf="center">...and how much did you heal?</ParagraphWS>
-            <RedBox alignSelf="center" width="150px" align="center" justify="between" pad="24px">
-              <FormField>
-                <TextInput
-                  type="number"
-                  value={harm}
-                  size="xlarge"
-                  textAlign="center"
-                  onChange={(e) => setHarm(parseInt(e.target.value))}
-                />
-              </FormField>
-            </RedBox>
-          </Box>
+        <Box fill align="start" justify="start">
+          <ParagraphWS alignSelf="start">Who did you have sex with?</ParagraphWS>
+          <Select
+            id="target-character-input"
+            aria-label="target-character-input"
+            name="target-character"
+            placeholder="Who?"
+            options={characters}
+            labelKey={'name'}
+            valueKey={'id'}
+            onChange={(e) => setotherCharacterId(e.value.id)}
+          />
         </Box>
         <Box fill="horizontal" direction="row" justify="end" gap="small">
           <ButtonWS
@@ -121,8 +100,8 @@ const HealHarmDialog: FC<HealHarmDialogProps> = ({ move, handleClose }) => {
           <ButtonWS
             label="APPLY"
             primary
-            onClick={() => !performingHealHarmMove && harm > 0 && !!otherCharacterId && handleHealHarmMove()}
-            disabled={performingHealHarmMove || harm === 0 || !otherCharacterId}
+            onClick={() => !performingAngelSpecialMove && !!otherCharacterId && handleAngelSpecialMove()}
+            disabled={performingAngelSpecialMove || !otherCharacterId}
           />
         </Box>
       </Box>
@@ -130,4 +109,4 @@ const HealHarmDialog: FC<HealHarmDialogProps> = ({ move, handleClose }) => {
   );
 };
 
-export default HealHarmDialog;
+export default AngelSpecialDialog;
