+++ src/components/dialogs/BarterDialog.tsx
@@ -1,26 +1,23 @@
 import React, { FC, useState } from 'react';
-import { useMutation } from '@apollo/client';
 import { useParams } from 'react-router-dom';
+import { useMutation } from '@apollo/client';
 import { Box, FormField, TextInput } from 'grommet';
 
-import DialogWrapper from './DialogWrapper';
-import { HeadingWS, ParagraphWS, RedBox, ButtonWS, makeWantKnownBackground } from '../config/grommetConfig';
-import PERFORM_MAKE_WANT_KNOWN_MOVE, {
-  PerformMakeWantKnownMoveData,
-  PerformMakeWantKnownMoveVars,
-} from '../mutations/performMakeWantKnownMove';
-import { Move, CharacterMove } from '../@types/staticDataInterfaces';
-import { useFonts } from '../contexts/fontContext';
-import { useGame } from '../contexts/gameContext';
+import DialogWrapper from '../DialogWrapper';
+import { HeadingWS, ParagraphWS, ButtonWS, barterBackground, RedBox } from '../../config/grommetConfig';
+import PERFORM_BARTER_MOVE, { PerformBarterMoveData, PerformBarterMoveVars } from '../../mutations/performBarterMove';
+import { CharacterMove, Move } from '../../@types/staticDataInterfaces';
+import { useFonts } from '../../contexts/fontContext';
+import { useGame } from '../../contexts/gameContext';
 
-interface MakeWantKnownDialogProps {
+interface BarterDialogProps {
   move: Move | CharacterMove;
   handleClose: () => void;
 }
 
-const MakeWantKnownDialog: FC<MakeWantKnownDialogProps> = ({ move, handleClose }) => {
+const BarterDialog: FC<BarterDialogProps> = ({ move, handleClose }) => {
   // -------------------------------------------------- Component state ---------------------------------------------------- //
-  const [barter, setBarter] = useState(0);
+  const [barter, setBarter] = useState(move.name === 'LIFESTYLE AND GIGS' ? 0 : 1);
   // -------------------------------------------------- 3rd party hooks ---------------------------------------------------- //
   const { gameId } = useParams<{ gameId: string }>();
 
@@ -29,22 +26,21 @@ const MakeWantKnownDialog: FC<MakeWantKnownDialogProps> = ({ move, handleClose }
   const { userGameRole } = useGame();
 
   // ------------------------------------------------------ graphQL -------------------------------------------------------- //
-  const [performMakeWantKnownMove, { loading: performingMakeWantKnownMove }] = useMutation<
-    PerformMakeWantKnownMoveData,
-    PerformMakeWantKnownMoveVars
-  >(PERFORM_MAKE_WANT_KNOWN_MOVE);
+  const [performBarterMove, { loading: performingBarterMove }] = useMutation<PerformBarterMoveData, PerformBarterMoveVars>(
+    PERFORM_BARTER_MOVE
+  );
 
   // ------------------------------------------------- Component functions -------------------------------------------------- //
   const currentBarter = userGameRole?.characters[0].barter || 0;
 
-  const handleMakeWantKnownMove = (move: Move | CharacterMove, barter: number) => {
+  const handleBarterMove = (move: Move | CharacterMove, barter: number) => {
     if (currentBarter - barter < 0) {
       console.warn("You don't have enough barter");
       return;
     }
-    if (!!userGameRole && userGameRole.characters.length === 1 && !performingMakeWantKnownMove) {
+    if (!!userGameRole && userGameRole.characters.length === 1 && !performingBarterMove) {
       try {
-        performMakeWantKnownMove({
+        performBarterMove({
           variables: {
             gameId,
             gameroleId: userGameRole.id,
@@ -59,28 +55,42 @@ const MakeWantKnownDialog: FC<MakeWantKnownDialogProps> = ({ move, handleClose }
       }
     }
   };
+
+  const getText = () => {
+    switch (move.name) {
+      case 'LIFESTYLE AND GIGS':
+        return 'How much barter will you spend?';
+      case 'GIVE BARTER':
+        return 'Give 1-barter?';
+      default:
+    }
+  };
   // ------------------------------------------------------ Render -------------------------------------------------------- //
 
   return (
-    <DialogWrapper background={makeWantKnownBackground} handleClose={handleClose}>
+    <DialogWrapper background={barterBackground} handleClose={handleClose}>
       <Box gap="24px">
         <HeadingWS crustReady={crustReady} level={4} alignSelf="start">
           {move.name}
         </HeadingWS>
-
-        <ParagraphWS alignSelf="start">How much jingle will you drop?</ParagraphWS>
-        <RedBox alignSelf="center" width="150px" align="center" justify="between" pad="24px">
-          <FormField>
-            <TextInput
-              type="number"
-              value={barter}
-              size="xlarge"
-              textAlign="center"
-              onChange={(e) => setBarter(parseInt(e.target.value))}
-            />
-          </FormField>
-        </RedBox>
-
+        {move.name === 'GIVE BARTER' ? (
+          <ParagraphWS alignSelf="start">{getText()}</ParagraphWS>
+        ) : (
+          <>
+            <ParagraphWS alignSelf="start">{getText()}</ParagraphWS>
+            <RedBox alignSelf="center" width="150px" align="center" justify="between" pad="24px">
+              <FormField>
+                <TextInput
+                  type="number"
+                  value={barter}
+                  size="xlarge"
+                  textAlign="center"
+                  onChange={(e) => setBarter(parseInt(e.target.value))}
+                />
+              </FormField>
+            </RedBox>
+          </>
+        )}
         <ParagraphWS alignSelf="start">{`You currently have ${currentBarter} barter`}</ParagraphWS>
         <Box fill="horizontal" direction="row" justify="end" gap="small">
           <ButtonWS
@@ -92,10 +102,10 @@ const MakeWantKnownDialog: FC<MakeWantKnownDialogProps> = ({ move, handleClose }
             onClick={handleClose}
           />
           <ButtonWS
-            label={'DROP'}
+            label={move.name === 'LIFESTYLE AND GIGS' ? 'SPEND' : 'GIVE'}
             primary
-            onClick={() => !!barter && !performingMakeWantKnownMove && handleMakeWantKnownMove(move, barter)}
-            disabled={!barter || performingMakeWantKnownMove || barter > 3}
+            onClick={() => !!barter && !performingBarterMove && handleBarterMove(move, barter)}
+            disabled={!barter || performingBarterMove}
           />
         </Box>
       </Box>
@@ -103,4 +113,4 @@ const MakeWantKnownDialog: FC<MakeWantKnownDialogProps> = ({ move, handleClose }
   );
 };
 
-export default MakeWantKnownDialog;
+export default BarterDialog;
