+++ src/components/dialogs/MakeWantKnownDialog.tsx
@@ -1,23 +1,26 @@
 import React, { FC, useState } from 'react';
-import { useParams } from 'react-router-dom';
 import { useMutation } from '@apollo/client';
+import { useParams } from 'react-router-dom';
 import { Box, FormField, TextInput } from 'grommet';
 
-import DialogWrapper from './DialogWrapper';
-import { HeadingWS, ParagraphWS, ButtonWS, barterBackground, RedBox } from '../config/grommetConfig';
-import PERFORM_BARTER_MOVE, { PerformBarterMoveData, PerformBarterMoveVars } from '../mutations/performBarterMove';
-import { CharacterMove, Move } from '../@types/staticDataInterfaces';
-import { useFonts } from '../contexts/fontContext';
-import { useGame } from '../contexts/gameContext';
+import DialogWrapper from '../DialogWrapper';
+import { HeadingWS, ParagraphWS, RedBox, ButtonWS, makeWantKnownBackground } from '../../config/grommetConfig';
+import PERFORM_MAKE_WANT_KNOWN_MOVE, {
+  PerformMakeWantKnownMoveData,
+  PerformMakeWantKnownMoveVars,
+} from '../../mutations/performMakeWantKnownMove';
+import { Move, CharacterMove } from '../../@types/staticDataInterfaces';
+import { useFonts } from '../../contexts/fontContext';
+import { useGame } from '../../contexts/gameContext';
 
-interface BarterDialogProps {
+interface MakeWantKnownDialogProps {
   move: Move | CharacterMove;
   handleClose: () => void;
 }
 
-const BarterDialog: FC<BarterDialogProps> = ({ move, handleClose }) => {
+const MakeWantKnownDialog: FC<MakeWantKnownDialogProps> = ({ move, handleClose }) => {
   // -------------------------------------------------- Component state ---------------------------------------------------- //
-  const [barter, setBarter] = useState(move.name === 'LIFESTYLE AND GIGS' ? 0 : 1);
+  const [barter, setBarter] = useState(0);
   // -------------------------------------------------- 3rd party hooks ---------------------------------------------------- //
   const { gameId } = useParams<{ gameId: string }>();
 
@@ -26,21 +29,22 @@ const BarterDialog: FC<BarterDialogProps> = ({ move, handleClose }) => {
   const { userGameRole } = useGame();
 
   // ------------------------------------------------------ graphQL -------------------------------------------------------- //
-  const [performBarterMove, { loading: performingBarterMove }] = useMutation<PerformBarterMoveData, PerformBarterMoveVars>(
-    PERFORM_BARTER_MOVE
-  );
+  const [performMakeWantKnownMove, { loading: performingMakeWantKnownMove }] = useMutation<
+    PerformMakeWantKnownMoveData,
+    PerformMakeWantKnownMoveVars
+  >(PERFORM_MAKE_WANT_KNOWN_MOVE);
 
   // ------------------------------------------------- Component functions -------------------------------------------------- //
   const currentBarter = userGameRole?.characters[0].barter || 0;
 
-  const handleBarterMove = (move: Move | CharacterMove, barter: number) => {
+  const handleMakeWantKnownMove = (move: Move | CharacterMove, barter: number) => {
     if (currentBarter - barter < 0) {
       console.warn("You don't have enough barter");
       return;
     }
-    if (!!userGameRole && userGameRole.characters.length === 1 && !performingBarterMove) {
+    if (!!userGameRole && userGameRole.characters.length === 1 && !performingMakeWantKnownMove) {
       try {
-        performBarterMove({
+        performMakeWantKnownMove({
           variables: {
             gameId,
             gameroleId: userGameRole.id,
@@ -55,42 +59,28 @@ const BarterDialog: FC<BarterDialogProps> = ({ move, handleClose }) => {
       }
     }
   };
-
-  const getText = () => {
-    switch (move.name) {
-      case 'LIFESTYLE AND GIGS':
-        return 'How much barter will you spend?';
-      case 'GIVE BARTER':
-        return 'Give 1-barter?';
-      default:
-    }
-  };
   // ------------------------------------------------------ Render -------------------------------------------------------- //
 
   return (
-    <DialogWrapper background={barterBackground} handleClose={handleClose}>
+    <DialogWrapper background={makeWantKnownBackground} handleClose={handleClose}>
       <Box gap="24px">
         <HeadingWS crustReady={crustReady} level={4} alignSelf="start">
           {move.name}
         </HeadingWS>
-        {move.name === 'GIVE BARTER' ? (
-          <ParagraphWS alignSelf="start">{getText()}</ParagraphWS>
-        ) : (
-          <>
-            <ParagraphWS alignSelf="start">{getText()}</ParagraphWS>
-            <RedBox alignSelf="center" width="150px" align="center" justify="between" pad="24px">
-              <FormField>
-                <TextInput
-                  type="number"
-                  value={barter}
-                  size="xlarge"
-                  textAlign="center"
-                  onChange={(e) => setBarter(parseInt(e.target.value))}
-                />
-              </FormField>
-            </RedBox>
-          </>
-        )}
+
+        <ParagraphWS alignSelf="start">How much jingle will you drop?</ParagraphWS>
+        <RedBox alignSelf="center" width="150px" align="center" justify="between" pad="24px">
+          <FormField>
+            <TextInput
+              type="number"
+              value={barter}
+              size="xlarge"
+              textAlign="center"
+              onChange={(e) => setBarter(parseInt(e.target.value))}
+            />
+          </FormField>
+        </RedBox>
+
         <ParagraphWS alignSelf="start">{`You currently have ${currentBarter} barter`}</ParagraphWS>
         <Box fill="horizontal" direction="row" justify="end" gap="small">
           <ButtonWS
@@ -102,10 +92,10 @@ const BarterDialog: FC<BarterDialogProps> = ({ move, handleClose }) => {
             onClick={handleClose}
           />
           <ButtonWS
-            label={move.name === 'LIFESTYLE AND GIGS' ? 'SPEND' : 'GIVE'}
+            label={'DROP'}
             primary
-            onClick={() => !!barter && !performingBarterMove && handleBarterMove(move, barter)}
-            disabled={!barter || performingBarterMove}
+            onClick={() => !!barter && !performingMakeWantKnownMove && handleMakeWantKnownMove(move, barter)}
+            disabled={!barter || performingMakeWantKnownMove || barter > 3}
           />
         </Box>
       </Box>
@@ -113,4 +103,4 @@ const BarterDialog: FC<BarterDialogProps> = ({ move, handleClose }) => {
   );
 };
 
-export default BarterDialog;
+export default MakeWantKnownDialog;
